cpp: move some code to AstExtensions
